챌린지, 배지, 유저챌린지 패키지 정리

+++ backend/src/main/java/com/runboo/domain/badge/dto/BadgeDto.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class BadgeDto {
-    private int badgeId;
+    private Long badgeId;
     private String name;
     private String description;
     private String iconUrl;
